Implemented general text post editor components for article and dopys creation/updation to avoid duplication
