+++ src/lib/components/Editor/Editors/VideoEditor.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import GeneralEditor from '../GeneralEditor';
+import { GeneralEditor } from '../GeneralEditor';
 import VideoEditorToolbar from './VideoEditorToolbar';
 
 interface IVideoEditorProps {
